+++ CotEditor/Sources/EditorTextView.swift
@@ -1365,7 +1365,7 @@ class EditorTextView: NSTextView, Themable, CurrentLineHighlighting, MultiCursor
                 indicator.color = self.insertionPointColor
             }
         }
-        self.selectedTextAttributes[.backgroundColor] = theme.effectiveSelectionColor
+        self.selectedTextAttributes[.backgroundColor] = theme.effectiveSelectionColor(for: self.effectiveAppearance)
         (self.layoutManager as? LayoutManager)?.invisiblesColor = theme.invisibles.color
         
         (self.window as? DocumentWindow)?.contentBackgroundColor = theme.background.color